src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
anyio==3.6.1
    # via watchfiles
asgiref==3.5.2
    # via django
attrs==22.2.0
    # via jsonschema
billiard==3.6.4.0
    # via celery
boto3==1.26.44
    # via -r requirements.in
botocore==1.29.44
    # via
    #   boto3
    #   s3transfer
celery==5.2.7
    # via -r requirements.in
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   uvicorn
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django==4.1
    # via
    #   -r requirements.in
    #   django-debug-toolbar
    #   django-storages
    #   djangorestframework
    #   drf-yasg
django-debug-toolbar==3.2
    # via -r requirements.in
django-redis-cache==3.0.0
    # via -r requirements.in
django-storages==1.13.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg[validation]==1.21.4
    # via -r requirements.in
gunicorn==20.0.4
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
httptools==0.4.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
importlib-metadata==6.0.0
    # via markdown
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.17.3
    # via swagger-spec-validator
kombu==5.2.4
    # via celery
markdown==3.4.1
    # via
    #   -r requirements.in
    #   markdown-urlize
markdown-urlize==0.2.0
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
numpy==1.24.1
    # via pandas
packaging==23.0
    # via drf-yasg
pandas==1.5.2
    # via -r requirements.in
prompt-toolkit==3.0.30
    # via click-repl
psycopg2==2.9.3
    # via -r requirements.in
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   botocore
    #   pandas
python-decouple==3.4
    # via -r requirements.in
python-dotenv==0.20.0
    # via uvicorn
pytz==2022.1
    # via
    #   celery
    #   djangorestframework
    #   drf-yasg
    #   pandas
pyyaml==6.0
    # via
    #   swagger-spec-validator
    #   uvicorn
redis==3.5.3
    # via
    #   -r requirements.in
    #   django-redis-cache
requests==2.28.1
    # via coreapi
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
s3transfer==0.6.0
    # via boto3
sentry-sdk==0.19.5
    # via -r requirements.in
six==1.15.0
    # via
    #   click-repl
    #   python-dateutil
sniffio==1.2.0
    # via anyio
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
swagger-spec-validator==3.0.3
    # via drf-yasg
typing-extensions==4.4.0
    # via swagger-spec-validator
tzdata==2022.7
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.2
    # via
    #   botocore
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.18.2
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.16.1
    # via uvicorn
watchgod==0.6
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
websockets==10.3
    # via uvicorn
whitenoise==5.2.0
    # via -r requirements.in
zipp==3.11.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
